Calendar: The delete events option status changed after rotate screen
Calendar through LoadManager to query the total number of events, and
update the delete events option status in onLoadFinished method. But
this method will be called at the first time load data or data have
changed, so can't update option status when rotating screen.

Calls the initLoader function in onCreate and onResume ensure that
the onLoadFinished can be callback.

CRs-Fixed: 650667

Change-Id: Id96f28bf648de374fb5d2cb7e0827e072f5342bd